move env utils
